move funcmap to its own sub package
