+++ pkg/compare/compare.go
@@ -17,6 +17,7 @@ import (
 
 	jsonpatch "github.com/evanphx/json-patch"
 	"github.com/gosimple/slug"
+	"github.com/openshift/kube-compare/pkg/funcmap"
 	"github.com/openshift/kube-compare/pkg/groups"
 	"github.com/samber/lo"
 	"github.com/spf13/cobra"
@@ -638,7 +639,7 @@ No CRs are unmatched to reference CRs
 {{- end }}
 `
 	var buf bytes.Buffer
-	tmpl, _ := template.New("Summary").Funcs(template.FuncMap{"toYaml": toYAML}).Parse(t)
+	tmpl, _ := template.New("Summary").Funcs(template.FuncMap{"toYaml": funcmap.ToYAML}).Parse(t)
 	_ = tmpl.Execute(&buf, s)
 	return strings.TrimSpace(buf.String())
 }
